% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.att_ipw_cox_fit}
\alias{print.att_ipw_cox_fit}
\title{Print an ATT-IPW Cox comparison}
\usage{
\method{print}{att_ipw_cox_fit}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link[=fit_att_ipw_cox]{fit_att_ipw_cox()}}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an ATT-IPW Cox comparison
}
