% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.mec_cox_fit}
\alias{print.mec_cox_fit}
\title{Print a MEC-Cox comparison}
\usage{
\method{print}{mec_cox_fit}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link[=fit_mec_cox]{fit_mec_cox()}}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a MEC-Cox comparison
}
