% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.unweighted_cox_fit}
\alias{print.unweighted_cox_fit}
\title{Print an unweighted Cox comparison}
\usage{
\method{print}{unweighted_cox_fit}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link[=fit_unweighted_cox]{fit_unweighted_cox()}}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an unweighted Cox comparison
}
